Moved IntlHead to separate component
